backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
azure-cosmos==4.5.1
pydantic==2.5.0
pydantic[email]
aiosmtplib==3.0.1
email-validator==2.1.0 